Per-core SCB layout
